+++ crates/python/src/lib.rs
@@ -11,6 +11,8 @@ pub mod qpu;
 pub mod qvm;
 pub mod register_data;
 
+pub(crate) mod py_sync;
+
 create_init_submodule! {
     classes: [
         execution_data::PyExecutionData,
@@ -49,38 +51,4 @@ create_init_submodule! {
 #[pymodule]
 fn qcs_sdk(py: Python<'_>, m: &PyModule) -> PyResult<()> {
     init_submodule("qcs_sdk", py, m)
-}
-
-pub(crate) mod utils {
-    /// Spawn and block on a future using the pyo3 tokio runtime.
-    /// Useful for returning a synchronous `PyResult`.
-    ///
-    ///
-    /// When used like the following:
-    /// ```rs
-    /// async fn say_hello(name: String) -> String {
-    ///     format!("hello {name}")
-    /// }
-    ///
-    /// #[pyo3(name="say_hello")]
-    /// pub fn py_say_hello(name: String) -> PyResult<String> {
-    ///     py_sync!(say_hello(name))
-    /// }
-    /// ```
-    ///
-    /// Becomes the associated "synchronous" python call:
-    /// ```py
-    /// assert say_hello("Rigetti") == "hello Rigetti"
-    /// ```
-    macro_rules! py_sync {
-        ($body: expr) => {{
-            let runtime = pyo3_asyncio::tokio::get_runtime();
-            let handle = runtime.spawn($body);
-            runtime
-                .block_on(handle)
-                .map_err(|err| PyRuntimeError::new_err(err.to_string()))?
-        }};
-    }
-
-    pub(crate) use py_sync;
 }
